requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
dominate==2.9.1
Flask==3.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.14.1
python-dotenv==1.1.1
requests==2.32.5
retrying==1.4.2
setuptools==80.9.0
tika==3.1.0
urllib3==2.5.0
visitor==0.1.3
Werkzeug==3.1.3
